+++ packages/react-wildcat-handoff/src/simple.js
@@ -1,9 +1,10 @@
 const simpleRender = require("./utils/simpleRender.js");
 
 function render(cfg) {
-    return function simpleHandoff(request, cookies, wildcatConfig) {
+    return function simpleHandoff(request, response, cookies, wildcatConfig) {
         cfg = Object.assign({}, cfg, {
             request,
+            response,
             wildcatConfig
         });